+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### Corrections
 
-- Added `\Wizaplace\SDK\User\UserService::UpdateuserAddresses`
+- Added `\Wizaplace\SDK\User\UserService::UpdateuserAddresses` Fixed a issue when updating addresses from full fields to empty fields and from full fields to new full fields
 - Added `\Wizaplace\SDK\Pim\Product::STANDBY`
 
 ## 1.35.0