+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### Corrections
 
-- Fixed an issue in `SDK\User\UserService::updateUserAddresses` that does not permit updating user addresses from full fields to empty fields and from full fields to new full fields
+- Fixed an issue in `SDK\User\UserService::updateUserAddresses` where addresses fields would not be updated
 - Added `\Wizaplace\SDK\Pim\Product::STANDBY`
 
 ## 1.35.0