+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- Fixed an issue in `SDK\User\UserService::updateUserAddresses` where addresses fields would not be updated
 - Added optionals parameters `$billing` and `$shipping` to `\Wizaplace\SDK\Organisation\OrganisationService::register`
 
 ## 1.36.0
@@ -13,7 +14,6 @@
 
 ### Corrections
 
-- Fixed an issue in `SDK\User\UserService::updateUserAddresses` where addresses fields would not be updated
 - Added `\Wizaplace\SDK\Pim\Product::STANDBY`
 
 ## 1.35.0