+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Corrections
 
+- Added `\Wizaplace\SDK\User\UserService::UpdateuserAddresses`
 - Added `\Wizaplace\SDK\Pim\Product::STANDBY`
 
 ## 1.35.0