+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## Next release
 
+### New features
+
 - Added optionals parameters `$billing` and `$shipping` to `\Wizaplace\SDK\Organisation\OrganisationService::register`
+
+### Corrections
+
+- Fixed an issue in `SDK\User\UserService::updateUserAddresses` where addresses fields would not be updated
 
 ## 1.36.0
 
@@ -13,7 +19,6 @@
 
 ### Corrections
 
-- Fixed an issue in `SDK\User\UserService::updateUserAddresses` where addresses fields would not be updated
 - Added `\Wizaplace\SDK\Pim\Product::STANDBY`
 
 ## 1.35.0