+++ CHANGELOG.md
@@ -2,14 +2,16 @@
 
 ## Next release
 
+### 1.37.0
+
 ### New features
 
-- Added optionals parameters `$billing` and `$shipping` to `\Wizaplace\SDK\Organisation\OrganisationService::register`
-- Added `SDK\Order\Order::getCompanyName` 
+- Added optional parameters `$billing` and `$shipping` to `\Wizaplace\SDK\Organisation\OrganisationService::register`
+- Added `\Wizaplace\SDK\Order\Order::getCompanyName` 
 
 ### Corrections
 
-- Fixed an issue in `SDK\User\UserService::updateUserAddresses` where addresses fields would not be updated
+- Fixed an issue in `\Wizaplace\SDK\User\UserService::updateUserAddresses` where address fields would not be updated
 
 ## 1.36.0